try to fix deadlock
the deadlock appears when the underlying c lib calls the callback
function into the rust code, meaning the deadlock originates from the
rust code, even though the mutex that isn't free'd is in the c lib.
in the callback, the rust code locks a mutex on receiving a message,
puts the message inside the mutex, and notifies a condvar that a message
has been given. until the message has been processed and the necessary
condvar has been set, the callback blocks on the same mutex which has
the side effect of not freeing the mutex of the underlying c lib, making
it impossible to send and receive new messages.
this means that, most likely, a condvar (`processed`) in the rust code
isn't called when it should, resulting in the callback function never
returning, which results in the client timing out on sending a new
message.

the current behaviour of _having to_ process a message before being
able to send a new one is not desirable, and a different interface
should be designed.
channels via `mpsc` come to mind, sending a message into the channel
when it has been received (which is mostly non-blocking) so that the
client can still be used even if the message is not read, but this is
also suboptimal as the channel _can_ block when it is full, resulting in
the same scenario as described above.